requirements.txt
PyGithub
docopt